refactor: utilizando o eyecomponent na pagina de login
